Start writing to unit tests (finally!)
